pyproject.toml
[project]
authors = [
  {name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com"},
]
dependencies = [
  "langgraph>=1.0.0",
  "langchain-google-genai>=0.2.0",
  "langsmith>=0.1.70",
  "python-dotenv>=1.0.1",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "psycopg2-binary>=2.9.9",
  "python-multipart>=0.0.6",
]
description = "Starter template for making a new agent LangGraph."
license = {text = "MIT"}
name = "agent"
readme = "README.md"
requires-python = ">=3.10"
version = "0.0.1"

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=73.0.0", "wheel"]

[tool.setuptools]
packages = [
  "src",
  "src.agent",
  "src.api",
  "src.api.routes",
  "src.api.middleware",
  "src.api.utils",
  "src.database",
  "src.tools",
  "src.observability",
  "src.utilities",
]
[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.ignore = [
  "UP006",
  "UP007",
  "UP035",
  "D417",
  "E501", # Line too long
  "D100", # Missing module-level docstring
]
lint.select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "D", # pydocstyle
  "D401", # First line should be in imperative mood
  "T201",
  "UP",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
  "anyio>=4.7.0",
  "langgraph-cli[inmem]>=0.4.14",
  "mypy>=1.13.0",
  "pytest>=8.3.5",
  "ruff>=0.8.2",
]
